Applying changes from GL to Gift transaction deletion

+++ csharp/ICT/Petra/Server/lib/MFinance/Gift/Gift.Transactions.cs
@@ -714,41 +714,61 @@ namespace Ict.Petra.Server.MFinance.Gift.WebConnectors
 
             if (AllValidationsOK)
             {
-                SubmissionResult = GiftBatchTDSAccess.SubmitChanges(AInspectDS, out AVerificationResult);
+            	SubmissionResult = GiftBatchTDSAccess.SubmitChanges(AInspectDS, out AVerificationResult);
 
                 if ((SubmissionResult == TSubmitChangesResult.scrOK) && AInspectDS.Tables.Contains("AGiftBatch")
-                    && AInspectDS.Tables.Contains("AGift"))
+                    && AInspectDS.Tables.Contains("AGift") && AInspectDS.AGift.Count > 0)
                 {
-                    DataRow[] foundGiftForDeletion = AInspectDS.AGift.Select(String.Format("{0} = '{1}'",
-                            AGiftTable.GetGiftStatusDBName(),
-                            MFinanceConstants.MARKED_FOR_DELETION));
-
-                    if (foundGiftForDeletion.Length == 1)
-                    {
-                        //A gift has been deleted
-                        //Accept the deletion of the single details row
-                        AInspectDS.AGiftDetail.AcceptChanges();
-                        AInspectDS.AGift.AcceptChanges();
-
-                        AGiftBatchTable clientBatchTable = (AGiftBatchTable)AInspectDS.AGiftBatch;
-                        AGiftBatchRow clientBatchRow = (AGiftBatchRow)clientBatchTable.Rows[0];
-
-                        AGiftRow giftRowClient = (AGiftRow)foundGiftForDeletion[0];
-
-                        giftRowClient.Delete();
-
-                        clientBatchRow.LastGiftNumber--;
-
-                        SubmissionResult = GiftBatchTDSAccess.SubmitChanges(AInspectDS, out AVerificationResult);
-
-                        //Accept the deletion of the single details row
-                        AInspectDS.AGiftDetail.AcceptChanges();
-                        AInspectDS.AGift.AcceptChanges();
-                        AInspectDS.AGiftBatch.AcceptChanges();
-                    }
-
-                    // Problem: unchanged rows will not arrive here? check after committing, and update the gift batch again
-                    // TODO: calculate hash of saved batch or batch of saved gift
+	            	if (AInspectDS.Tables.Contains("AGiftDetail"))
+	                {
+		            	AInspectDS.AGiftDetail.AcceptChanges();
+	                }
+	
+	                AInspectDS.AGift.AcceptChanges();
+	            	
+	            	AGiftRow tranR = (AGiftRow)AInspectDS.AGift.Rows[0];
+	
+	                Int32 currentLedger = tranR.LedgerNumber;
+	                Int32 currentBatch = tranR.BatchNumber;
+	                Int32 giftToDelete = 0;
+	
+	                try
+	                {
+	                    DataRow[] foundGiftsForDeletion = AInspectDS.AGift.Select(String.Format("{0} = '{1}'",
+	                            AGiftTable.GetGiftStatusDBName(),
+	                            MFinanceConstants.MARKED_FOR_DELETION));
+	
+	                    if (foundGiftsForDeletion.Length > 0)
+	                    {
+	                        AGiftRow giftRowClient = null;
+	                        
+	                        for (int i = 0; i < foundGiftsForDeletion.Length; i++)
+	                        {
+		                    	//A gift has been deleted
+		                        giftRowClient = (AGiftRow)foundGiftsForDeletion[i];
+		                        
+		                        giftToDelete = giftRowClient.GiftTransactionNumber;
+	                            TLogging.Log(String.Format("Gift to Delete: {0} from Batch: {1}",
+	                                    giftToDelete,
+	                                    currentBatch));
+		
+		                        giftRowClient.Delete();
+	                        }
+	                    }
+	
+	                    SubmissionResult = GiftBatchTDSAccess.SubmitChanges(AInspectDS, out AVerificationResult);
+	                }
+	                catch (Exception ex)
+	                {
+	                	TLogging.Log("Saving DataSet: " + ex.Message);
+	                	
+	                	TLogging.Log(String.Format("Error trying to save changes: {0} in Batch: {1}",
+	                            giftToDelete,
+	                            currentBatch
+	                            ));
+	                	
+	                	SubmissionResult = TSubmitChangesResult.scrError;
+	                }
                 }
             }
